server -> backend, client -> frontend
