+++ backend/.sequelizerc
@@ -1,0 +1,9 @@
+'use strict';
+const path = require('path');
+
+module.exports = {
+  'config': path.resolve('./database/config', 'config.js'),
+  'models-path': path.resolve('./database/models'),
+  'seeders-path': path.resolve('./database/seeders'),
+  'migrations-path': path.resolve('./database/migrations'),
+};
