+++ backend/routes/activities.js
