+++ frontend/src/root.css
